backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.1.2
werkzeug==3.0.1
sqlalchemy==2.0.23
pyjwt==2.8.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dateutil==2.8.2
requests==2.31.0
